Move ImageCommentsUIComposer to production
